seperating roots and controller/ removing redundant files
