+++ package.json
@@ -18,6 +18,7 @@
     "ejs-mate": "^4.0.0",
     "express": "^4.18.2",
     "express-session": "^1.18.0",
+    "helmet": "^7.1.0",
     "mate": "^0.21.0",
     "moment": "^2.30.1",
     "morgan": "^1.10.0",
