--- routes/users.js
@@ -1,1 +1,0 @@
-//routes relating to user specific action
