--- views/dailylog/chart.ejs
@@ -1,51 +1,0 @@
-<% layout('layouts/boilerplate') %>
-<script src="https://cdnjs.cloudflare.com/ajax/libs/Chart.js/4.4.1/chart.umd.js" integrity="sha512-ZwR1/gSZM3ai6vCdI+LVF1zSq/5HznD3ZSTk7kajkaj4D292NLuduDCO1c/NT8Id+jE58KYLKT7hXnbtryGmMg==" crossorigin="anonymous" referrerpolicy="no-referrer"></script>
-
-<div>
-
-    
-            <!-- Chart canvas will be dynamically created here -->
-            <canvas id="emotionChart"></canvas>
-        </div>
-        <script>
-       
-            console.log("WHY ISNT THIS LOGGING")
-            var yvals = [1,2,3,4,5,6,7,8];
-            var xlabels = [1, 2, 3, 4, 5 ,6 ,7 ,8];
-            const chartdata= {
-                labels: xlabels,
-                datasets:[{
-                    data:yvals,
-                    lineTension:0,
-                    pointRadius:3,
-                    backgroundColor:'lightgray',
-                    bordercColor:'red',
-                    borderWidth:1
-                }]
-            };
-        
-            const chartconfig = {
-                type:'line',
-                data: chartdata,
-                options:{
-                    animation: true,
-                    maintainAspectRatio:true,
-                    scales:{y:{min:0, max : 20}},
-                    plugins:{
-                        legend:{display:false},
-                        tooltips:{enabled:true},
-                        tile:{
-                            display:true,
-                            text: "Emotions over time",
-                            font:{size:20},
-                        }
-                    }
-                }
-        
-            };
-        
-            const chart = new Chart('emotionChart',chartconfig);
-        
-        
-        </script>
-
