+++ views/dailylog/show.ejs
@@ -2,7 +2,7 @@
     <script src="https://cdnjs.cloudflare.com/ajax/libs/Chart.js/4.4.1/chart.umd.js"
         integrity="sha512-ZwR1/gSZM3ai6vCdI+LVF1zSq/5HznD3ZSTk7kajkaj4D292NLuduDCO1c/NT8Id+jE58KYLKT7hXnbtryGmMg=="
         crossorigin="anonymous" referrerpolicy="no-referrer"></script>
-    <script src="../../public/javascripts/chartjscript.js"> </script>
+   
     <div class="container-fluid">
         <% if (loggedin) { %>
             <h2 class="mt-4 mb-3">
@@ -72,7 +72,7 @@
                                                                                         <div
                                                                                             class="col-md-6 text-right">
                                                                                             <a class="btn btn-secondary"
-                                                                                                href="/view/<%= snapshotid %>">View
+                                                                                                href="/dailylog/view/<%= snapshotid %>">View
                                                                                                 Record</a>
                                                                                         </div>
                                                                                     </div>
@@ -101,12 +101,12 @@
                                                                                         <div
                                                                                             class="col-md-6 text-right">
                                                                                             <a class="btn btn-info"
-                                                                                                href="/edit/<%= snapshotid %>">Edit</a>
+                                                                                                href="/dailylog/edit/<%= snapshotid %>">Edit</a>
                                                                                             <form method="post"
                                                                                                 class="d-inline">
                                                                                                 <button
                                                                                                     class="btn btn-danger"
-                                                                                                    formaction="/del/<%= snapshotid %>">Delete</button>
+                                                                                                    formaction="/dailylog/del/<%= snapshotid %>">Delete</button>
                                                                                             </form>
                                                                                         </div>
                                                                                     </div>
