+++ views/dailylog/view.ejs
@@ -45,10 +45,10 @@
                                     </p>
                                     <div class="col-md-12 text-right">
                                         <a class="card-link btn btn-secondary" href="/dailylog/showall">Back</a>
-                                        <a class="card-link btn btn-info" href="/edit/<%= snapshotid %>">Edit</a>
+                                        <a class="card-link btn btn-info" href="/dailylog/edit/<%= snapshotid %>">Edit</a>
                                         <form method="post" class="d-inline">
                                             <button class="d-inline btn btn-danger"
-                                                formaction="/del/<%= snapshotid %>">Delete</button>
+                                                formaction="/dailylog/del/<%= snapshotid %>">Delete</button>
                                         </form>
                                     </div>
                                     <br>
